packages/plugin-camelai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neocortex-camelai"
version = "0.1.0"
description = "CAMEL AI integrations for TinyHumans memory API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "tinyhumansai>=0.5",
    "camel-ai>=0.1.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]
examples = ["python-dotenv>=1.0"]

[tool.setuptools.packages.find]
include = ["neocortex_camelai*"]

[tool.ruff]
line-length = 88
target-version = "py39"
